JsRuntime::mod_new -> ModuleMap::new_module
